pyproject.toml
[project]
name = "chicago-bike-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.41.0",
    "ipykernel>=7.2.0",
    "pandas>=3.0.2",
]
